ai/requirements.txt
flask
numpy
pandas
scikit-learn
nltk

pymongo
python-dotenv
joblib